忘记了变量判断导致解析出来的权限标识符缺少前面的0字符

+++ src/main/java/com/ztianzeng/jjvm/ConstantInfo.java
@@ -91,7 +91,7 @@ public class ConstantInfo {
     static class ConstantLongInfo extends ConstantInfo {
         long val;
 
-        ConstantLongInfo(byte high, byte low) {
+        ConstantLongInfo(int high, int low) {
             this.val = ((long) high << 32) + low;
         }
     }
@@ -100,7 +100,7 @@ public class ConstantInfo {
         double val;
 
         // TODO: 2017/9/13 还需要计算
-        ConstantDoubleInfo(byte high, byte low) {
+        ConstantDoubleInfo(int high, int low) {
             this.val = (high << 32) + low;
         }
     }
